Prestcon API/requirements.txt
certifi==2018.1.18
chardet==3.0.4
Django==1.11.7
django-cors-headers==2.2.0
django-easy-pdf==0.1.1
django-filter==1.1.0
django-filters==0.2.1
django-multiselectfield==0.1.8
django-rest-framework==0.1.0
djangorestframework==3.4.6
djangorestframework-expiring-authtoken==0.1.4
html5lib==1.0.1
httplib2==0.11.3
idna==2.6
Pillow==5.1.0
PyPDF2==1.26.0
pytz==2018.4
reportlab==3.4.0
requests==2.18.4
six==1.11.0
urllib3==1.22
webencodings==0.5.1
xhtml2pdf==0.2.2
